+++ pegtl/internal/pegtl_string.hh
@@ -55,10 +55,10 @@ namespace PEGTL_NAMESPACE
 #define PEGTL_INTERNAL_EXPAND(...) __VA_ARGS__
 
 #define PEGTL_INTERNAL_STRING_AT( S, x, n ) \
-   pegtl::internal::string_at< S, ( 0##n < sizeof( x ) ) ? x[ 0##n ] : 0, ( 0##n < sizeof( x ) - 1 ) >::type
+   PEGTL_NAMESPACE::internal::string_at< S, ( 0##n < sizeof( x ) ) ? x[ 0##n ] : 0, ( 0##n < sizeof( x ) - 1 ) >::type
 
 #define PEGTL_INTERNAL_JOIN_8( M, S, x, n ) \
-   pegtl::internal::string_join<            \
+   PEGTL_NAMESPACE::internal::string_join<            \
    PEGTL_INTERNAL_DEFER( M )( S, x, n##0 ), \
    PEGTL_INTERNAL_DEFER( M )( S, x, n##1 ), \
    PEGTL_INTERNAL_DEFER( M )( S, x, n##2 ), \
@@ -81,12 +81,12 @@ namespace PEGTL_NAMESPACE
    PEGTL_INTERNAL_EXPAND( \
       PEGTL_INTERNAL_EXPAND( \
          PEGTL_INTERNAL_EXPAND( \
-            pegtl::internal::string_max_length< PEGTL_INTERNAL_STRING_512( S, x, ), sizeof( x ) - 1 >::type ) ) )
+            PEGTL_NAMESPACE::internal::string_max_length< PEGTL_INTERNAL_STRING_512( S, x, ), sizeof( x ) - 1 >::type ) ) )
 
 #define pegtl_string_t( x ) \
-   PEGTL_INTERNAL_STRING( pegtl::ascii::string, x )
+   PEGTL_INTERNAL_STRING( PEGTL_NAMESPACE::ascii::string, x )
 
 #define pegtl_istring_t( x ) \
-   PEGTL_INTERNAL_STRING( pegtl::ascii::istring, x )
+   PEGTL_INTERNAL_STRING( PEGTL_NAMESPACE::ascii::istring, x )
 
 #endif